Switch mathoid fake endpoints to another controller
